insurance-app/services/user_service/app/v1/requirements.txt
fastapi
uvicorn
httpx
python-jose
itsdangerous
python-multipart
pyyaml
redis[asyncio]
